Moved Giphy to own module
